Some List->Vector conversions

+++ scene/resources/texture.cpp
@@ -697,7 +697,7 @@ Error StreamTexture::_load_data(se_string_view p_path, int &tw, int &th, int &tw
                 PoolVector<uint8_t>::Write w = img_data.write();
 
                 int ofs = 0;
-                for (int i = 0; i < mipmap_images.size(); i++) {
+                for (size_t i = 0; i < mipmap_images.size(); i++) {
 
                     PoolVector<uint8_t> id = mipmap_images[i]->get_data();
                     int len = id.size();
@@ -1363,7 +1363,7 @@ RID LargeTexture::get_rid() const {
 
 bool LargeTexture::has_alpha() const {
 
-    for (int i = 0; i < pieces.size(); i++) {
+    for (size_t i = 0; i < pieces.size(); i++) {
         if (pieces[i].texture->has_alpha())
             return true;
     }
@@ -1373,7 +1373,7 @@ bool LargeTexture::has_alpha() const {
 
 void LargeTexture::set_flags(uint32_t p_flags) {
 
-    for (int i = 0; i < pieces.size(); i++) {
+    for (size_t i = 0; i < pieces.size(); i++) {
         pieces[i].texture->set_flags(p_flags);
     }
 }
@@ -1424,7 +1424,7 @@ void LargeTexture::clear() {
 Array LargeTexture::_get_data() const {
 
     Array arr;
-    for (int i = 0; i < pieces.size(); i++) {
+    for (size_t i = 0; i < pieces.size(); i++) {
         arr.push_back(pieces[i].offset);
         arr.push_back(pieces[i].texture);
     }
@@ -1459,7 +1459,7 @@ Ref<Texture> LargeTexture::get_piece_texture(int p_idx) const {
 Ref<Image> LargeTexture::to_image() const {
 
     Ref<Image> img(make_ref_counted<Image>(this->get_width(), this->get_height(), false, Image::FORMAT_RGBA8));
-    for (int i = 0; i < pieces.size(); i++) {
+    for (size_t i = 0; i < pieces.size(); i++) {
 
         Ref<Image> src_img = pieces[i].texture->get_data();
         img->blit_rect(src_img, Rect2(0, 0, src_img->get_width(), src_img->get_height()), pieces[i].offset);
